Fix date handling re: run start date/time & nitrate/diatom data reduction.
SOG run start date/time is midnight on the day of the CTD cast used
for initialization, so adjust the value read from the infile to
reflect that.

When the nitrate concentration and diatom biomass results are reduced
to daily min and max values the date calculation must take leap-years
into account.